MatchActivityにあった関数をMatchPresenterに移植
